// be-ts-express/src/enums/state-field-type.enum.ts
export enum StateFieldType {
  Shipment = "shipment",
  Medal = "medal",
}
export enum StateFieldSubType {
  Full = "full",
  Coleta = "coleta",
  Correios = "correios",
  Others = "others",
  MedalLider = "medalLider",
  MedalGold = "medalGold",
  MedalPlatinum = "medalPlatinum",
  NoMedal = "noMedal",
}